Restructure the code to move the logic to internal packages
